+++ src/assets/components/Form/Form.sass
@@ -117,7 +117,7 @@ $InputFocus: #4c43a1
     &__btn-submit
         width: 100%
         height: calc(1.5em + 0.75rem + 2px)
-        margin-bottom: 5px
+        margin-bottom: 10px
         padding: 0.375rem 0.75rem
         color: #fff
         font: 400 normal 1rem/1.5 $Roboto
@@ -137,4 +137,7 @@ $InputFocus: #4c43a1
     &__error
         display: flex
         color: red,
+        font: 400 normal .8rem/1.5 $Roboto
+    &__helper
+        color: $GrayText
         font: 400 normal .8rem/1.5 $Roboto
