+++ src/assets/redux/selectors/FormSelectors.ts
@@ -1,11 +1,11 @@
 import { RootState } from "../store";
 
 export const getInputFirstName = (state: RootState) =>
-    state.FormSlice.firstNameField;
+    state.form.firstNameField;
 export const getInputSecondName = (state: RootState) =>
-    state.FormSlice.secondNameField;
-export const getInputEmail = (state: RootState) => state.FormSlice.emailField;
-export const getCategory = (state: RootState) => state.FormSlice.category;
-export const getMessage = (state: RootState) => state.FormSlice.messageField;
-export const getImage = (state: RootState) => state.FormSlice.addImage;
-export const getValidation = (state: RootState) => state.FormSlice.isValid;
+    state.form.secondNameField;
+export const getInputEmail = (state: RootState) => state.form.emailField;
+export const getCategory = (state: RootState) => state.form.category;
+export const getMessage = (state: RootState) => state.form.messageField;
+export const getImage = (state: RootState) => state.form.addImage;
+export const getValidation = (state: RootState) => state.form.isValid;
